+++ src/index.tsx
@@ -4,12 +4,19 @@ import './styles/index.css';
 import './styles/todo-list.css';
 import './styles/filters.css';
 
-import { HashRouter } from 'react-router-dom';
+import { HashRouter, Route, Routes } from 'react-router-dom';
 import { App } from './App';
+import { TodoApp } from './components/TodoApp';
 
 ReactDOM.render(
   <HashRouter>
-    <App />
+    <Routes>
+      <Route path="/" element={<App />}>
+        <Route index element={<TodoApp />} />
+        <Route path="completed" element={<TodoApp />} />
+        <Route path="active" element={<TodoApp />} />
+      </Route>
+    </Routes>
   </HashRouter>,
   document.getElementById('root'),
 );
