+++ src/styles/index.css
@@ -107,6 +107,10 @@ body {
   box-shadow: inset 0 -2px 1px rgba(0,0,0,0.03);
 }
 
+.new-todo:focus {
+  outline: 0;
+}
+
 .main {
   position: relative;
   z-index: 2;
